Split mockData and handlers
