+++ services/orchest-webserver/client/src/contexts/__tests__/ProjectsContext.test.tsx
@@ -5,7 +5,7 @@ import {
   chance,
   getPipelineMedadatas,
   mockProjectCollection,
-} from "../../__mocks__/handlers.mock";
+} from "../../__mocks__/mockData.mock";
 import {
   ProjectsContextProvider,
   useProjectsContext,
